Furi: drop CMSIS_OS header and unused api, partially refactor and cleanup the rest
