+++ applications/rpc/rpc_app.c
@@ -1,4 +1,3 @@
-#include "cmsis_os2.h"
 #include "flipper.pb.h"
 #include <core/record.h>
 #include "rpc_i.h"
