+++ applications/storage_move_to_sd/storage_move_to_sd.c
@@ -1,5 +1,4 @@
 #include "storage_move_to_sd.h"
-#include "cmsis_os2.h"
 #include <core/common_defines.h>
 #include <core/log.h>
 #include "loader/loader.h"
