+++ applications/subghz/views/subghz_test_packet.c
@@ -206,7 +206,7 @@ void subghz_test_packet_enter(void* context) {
 
     furi_hal_subghz_start_async_rx(subghz_test_packet_rx_callback, instance);
 
-    osTimerStart(instance->timer, osKernelGetTickFreq() / 4);
+    osTimerStart(instance->timer, furi_kernel_get_tick_frequency() / 4);
 }
 
 void subghz_test_packet_exit(void* context) {
