+++ lib/freertos.scons
@@ -20,7 +20,6 @@ libenv.ApplyLibFlags()
 sources = libenv.Glob("FreeRTOS-Kernel/*.c", source=True)
 sources += [
     "FreeRTOS-Kernel/portable/GCC/ARM_CM4F/port.c",
-    "FreeRTOS-glue/cmsis_os2.c",
 ]
 
 lib = libenv.StaticLibrary("${FW_LIB_NAME}", sources)
